requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.7
Django==3.2.1
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
flake8==3.9.2
gunicorn==20.1.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-slugify==5.0.2
pytz==2021.1
reportlab==3.5.67
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.1
